Add unfinished version of level3 and level3 activity

+++ phase1/app/src/main/java/com/example/uoftlife/Score.java
@@ -5,14 +5,15 @@ class Score{
     private int score;
 
     // The difficulty of this Score
-    private byte difficulty;
+    // TODO: Remeber to change the type
+    private int difficulty;
 
     //The user that create this score
     //TODO: Remember to change the type
     private String user;
 
     //Score is 0 at first. Do not need to pass in score.
-    Score(byte difficulty, String user) {
+    Score(int difficulty, String user) {
         this.difficulty = difficulty;
         this.user = user;
     }
